Total song count in "All Songs" category, UI edits to toolbar
Updates the "All Songs" category in the category list to display
the total number of songs. It also adds a "No songs found."
message to the song list view when there are no songs to display.

Also includes minor UI adjustments to the song list view toolbar
and adds .cursor rules and gitignore entries.